# Conflicts:
#	src/main/java/cn/iyque/controller/IYcallbackController.java
#	src/main/java/cn/iyque/service/impl/IYqueDefaultMsgImpl.java

+++ src/main/java/cn/iyque/dao/IYqueConfigDao.java
@@ -1,8 +1,7 @@
 package cn.iyque.dao;
 
-import cn.iyque.domain.IYqueConfig;
-import cn.iyque.domain.IYqueDefaultMsg;
+import cn.iyque.entity.IYqueConfig;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IYqueConfigDao  extends JpaRepository<IYqueConfig,Integer> {
+public interface IYqueConfigDao  extends JpaRepository<IYqueConfig,Long> {
 }
