+++ src/main/java/cn/iyque/dao/IYqueDefaultMsgDao.java
@@ -1,8 +1,7 @@
 package cn.iyque.dao;
 
-import cn.iyque.domain.IYqueConfig;
-import cn.iyque.domain.IYqueDefaultMsg;
+import cn.iyque.entity.IYqueDefaultMsg;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IYqueDefaultMsgDao extends JpaRepository<IYqueDefaultMsg,Integer> {
+public interface IYqueDefaultMsgDao extends JpaRepository<IYqueDefaultMsg,Long> {
 }
